+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- GS plugin: Allow to select only one installation in the InstallationsSelector component.
+
 ## [0.24.0] - 2024-06-13
 
 ### Added