+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.25.0] - 2024-06-20
+
 ### Added
 
 - GS plugin: Add scaffolder custom field extension for picking a workload cluster.
@@ -521,7 +523,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Disable anonymous access.
 
-[Unreleased]: https://github.com/giantswarm/backstage/compare/v0.24.0...HEAD
+[Unreleased]: https://github.com/giantswarm/backstage/compare/v0.25.0...HEAD
+[0.25.0]: https://github.com/giantswarm/backstage/compare/v0.24.0...v0.25.0
 [0.24.0]: https://github.com/giantswarm/backstage/compare/v0.23.1...v0.24.0
 [0.23.1]: https://github.com/giantswarm/backstage/compare/v0.23.0...v0.23.1
 [0.23.0]: https://github.com/giantswarm/backstage/compare/v0.22.2...v0.23.0